requirements.txt
altgraph==0.17.4
iniconfig==2.0.0
ntplib==0.4.0
packaging==24.0
pluggy==1.5.0
pygame==2.5.2
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.5
pytest==8.1.1
PyYAML==6.0.1
